moved variables type in type file
